+++ augur/utils.py
@@ -11,13 +11,19 @@ from treetime.utils import numeric_date
 from collections import defaultdict
 from pkg_resources import resource_stream
 from io import TextIOWrapper
-from textwrap import dedent
 from .__version__ import __version__
 import packaging.version as packaging_version
 from .validate import validate, ValidateError, load_json_schema
 
+from augur.util_support.color_parser import ColorParser
+from augur.util_support.date_disambiguator import DateDisambiguator
+from augur.util_support.metadata_file import MetadataFile
+from augur.util_support.shell_command_runner import ShellCommandRunner
+
+
 class AugurException(Exception):
     pass
+
 
 @contextmanager
 def open_file(fname, mode):
@@ -62,78 +68,11 @@ def get_json_name(args, default=None):
             raise ValueError("Please specify a name for the JSON file containing the results.")
 
 
-def ambiguous_date_to_date_range(mydate, fmt, min_max_year=None):
-    from datetime import datetime
-    sep = fmt.split('%')[1][-1]
-    min_date, max_date = {}, {}
-    today = datetime.today().date()
-
-    for val, field  in zip(mydate.split(sep), fmt.split(sep+'%')):
-        f = 'year' if 'y' in field.lower() else ('day' if 'd' in field.lower() else 'month')
-        if 'XX' in val:
-            if f=='year':
-                if min_max_year:
-                    min_date[f]=min_max_year[0]
-                    if len(min_max_year)>1:
-                        max_date[f]=min_max_year[1]
-                    elif len(min_max_year)==1:
-                        max_date[f]=4000 #will be replaced by 'today' below.
-                else:
-                    return None, None
-            elif f=='month':
-                min_date[f]=1
-                max_date[f]=12
-            elif f=='day':
-                min_date[f]=1
-                max_date[f]=31
-        else:
-            min_date[f]=int(val)
-            max_date[f]=int(val)
-    max_date['day'] = min(max_date['day'], 31 if max_date['month'] in [1,3,5,7,8,10,12]
-                                           else 28 if max_date['month']==2 else 30)
-    lower_bound = datetime(year=min_date['year'], month=min_date['month'], day=min_date['day']).date()
-    upper_bound = datetime(year=max_date['year'], month=max_date['month'], day=max_date['day']).date()
-    return (lower_bound, upper_bound if upper_bound<today else today)
+def ambiguous_date_to_date_range(uncertain_date, fmt, min_max_year=None):
+    return DateDisambiguator(uncertain_date, fmt=fmt, min_max_year=min_max_year).range()
 
 def read_metadata(fname, query=None):
-    if not fname:
-        print("ERROR: read_metadata called without a filename")
-        return {}, []
-    if os.path.isfile(fname):
-        try:
-            metadata = pd.read_csv(fname, sep='\t' if fname[-3:]=='tsv' else ',',
-                                    skipinitialspace=True).fillna('')
-        except pd.errors.ParserError as e:
-            print("Error reading metadata file {}".format(fname))
-            print(e)
-            sys.exit(2)
-        if query:
-            try:
-                metadata.query(query, inplace=True)
-            except Exception as e:
-                # Would like to make this more specific, but Pandas throws multiple different
-                # errors from panda specific to python generic errors.
-                print("ERROR: Invalid query string: '{}'".format(query))
-                print(e)
-                sys.exit(2)
-        meta_dict = {}
-        for ii, val in metadata.iterrows():
-            if hasattr(val, "strain"):
-                if val.strain in meta_dict:
-                    raise ValueError("Duplicate strain '{}'".format(val.strain))
-                meta_dict[val.strain] = val.to_dict()
-            elif hasattr(val, "name"):
-                if val.name in meta_dict:
-                    raise ValueError("Duplicate name '{}'".format(val.name))
-                meta_dict[val.name] = val.to_dict()
-            else:
-                print("ERROR: meta data file needs 'name' or 'strain' column")
-
-        return meta_dict, list(metadata.columns)
-    else:
-        print("ERROR: meta data file ({}) does not exist".format(fname))
-        return {}, []
-
+    return MetadataFile(fname, query).read()
 
 def get_numerical_dates(meta_dict, name_col = None, date_col='date', fmt=None, min_max_year=None):
     if fmt:
@@ -438,46 +377,7 @@ def read_lat_longs(overrides=None, use_defaults=True):
     return coordinates
 
 def read_colors(overrides=None, use_defaults=True):
-    colors = {}
-    # TODO: make parsing of tsv files more robust while allow for whitespace delimiting for backwards compatibility
-    def add_line(line):
-        if line.startswith('#'):
-            return
-        fields = line.strip().split() if not '\t' in line else line.strip().split('\t')
-        if not fields:
-            return # blank lines
-        if len(fields) != 3:
-            print("WARNING: Color map file contains invalid line. Please make sure not to mix tabs and spaces as delimiters (use only tabs):",line)
-            return
-        trait, trait_value, hex_code = fields[0].lower(), fields[1].lower(), fields[2]
-        if not hex_code.startswith("#") or len(hex_code) != 7:
-            print("WARNING: Color map file contained this invalid hex code: ", hex_code)
-            return
-        # If was already added, delete entirely so order can change to order in user-specified file
-        # (even though dicts shouldn't be relied on to have order)
-        if (trait, trait_value) in colors:
-            del colors[(trait, trait_value)]
-        colors[(trait, trait_value)] = hex_code
-
-
-    if use_defaults:
-        with resource_stream(__package__, "data/colors.tsv") as stream:
-            with TextIOWrapper(stream, "utf-8") as defaults:
-                for line in defaults:
-                    add_line(line)
-
-    if overrides:
-        if os.path.isfile(overrides):
-            with open(overrides, encoding='utf-8') as fh:
-                for line in fh:
-                    add_line(line)
-        else:
-            print("WARNING: Couldn't open color definitions file {}.".format(overrides))
-    color_map = defaultdict(list)
-    for (trait, trait_value), hex_code in colors.items():
-        color_map[trait].append((trait_value, hex_code))
-
-    return color_map
+    return ColorParser(mapping_filename=overrides, use_defaults=use_defaults).mapping
 
 def write_VCF_translation(prot_dict, vcf_file_name, ref_file_name):
     """
@@ -567,7 +467,7 @@ def write_VCF_translation(prot_dict, vcf_file_name, ref_file_name):
 
 shquote = shlex.quote
 
-def run_shell_command(cmd, raise_errors = False, extra_env = None):
+def run_shell_command(cmd, raise_errors=False, extra_env=None):
     """
     Run the given command string via Bash with error checking.
 
@@ -578,66 +478,7 @@ def run_shell_command(cmd, raise_errors = False, extra_env = None):
     If an *extra_env* mapping is passed, the provided keys and values are
     overlayed onto the default subprocess environment.
     """
-    env = os.environ.copy()
-
-    if extra_env:
-        env.update(extra_env)
-
-    shargs = ['-c', "set -euo pipefail; " + cmd]
-
-    if os.name == 'posix':
-        shellexec = ['/bin/bash']
-    else:
-        # We try best effort on other systems. For now that means nt/java.
-        shellexec = ['env', 'bash']
-
-    try:
-        # Use check_call() instead of run() since the latter was added only in Python 3.5.
-        subprocess.check_output(
-            shellexec + shargs,
-            shell = False,
-            stderr = subprocess.STDOUT,
-            env = env)
-
-    except subprocess.CalledProcessError as error:
-        print_error(
-            "{out}\nshell exited {rc} when running: {cmd}{extra}",
-            out = error.output,
-            rc  = error.returncode,
-            cmd = cmd,
-            extra = "\nAre you sure this program is installed?" if error.returncode==127 else "",
-        )
-        if raise_errors:
-            raise
-        else:
-            return False
-
-    except FileNotFoundError as error:
-        print_error(
-            """
-            Unable to run shell commands using {shell}!
-
-            Augur requires {shell} to be installed.  Please open an issue on GitHub
-            <https://github.com/nextstrain/augur/issues/new> if you need assistance.
-            """,
-            shell = ' and '.join(shellexec)
-        )
-        if raise_errors:
-            raise
-        else:
-            return False
-
-    else:
-        return True
-
-
-def print_error(message, **kwargs):
-    """
-    Formats *message* with *kwargs* using :meth:`str.format` and
-    :func:`textwrap.dedent` and uses it to print an error message to
-    ``sys.stderr``.
-    """
-    print("\nERROR: " + dedent(message.format(**kwargs)).lstrip("\n")+"\n", file = sys.stderr)
+    return ShellCommandRunner(cmd, raise_errors=raise_errors, extra_env=extra_env).run()
 
 
 def first_line(text):
@@ -900,3 +741,8 @@ def load_mask_sites(mask_file):
         mask_sites = read_mask_file(mask_file)
     print("%d masking sites read from %s" % (len(mask_sites), mask_file))
     return mask_sites
+
+VALID_NUCLEOTIDES = { # http://reverse-complement.com/ambiguity.html
+    "A", "G", "C", "T", "U", "N", "R", "Y", "S", "W", "K", "M", "B", "V", "D", "H", "-",
+    "a", "g", "c", "t", "u", "n", "r", "y", "s", "w", "k", "m", "b", "v", "d", "h", "-"
+}