+++ scripts/diff_jsons.py
@@ -13,6 +13,7 @@ if __name__ == "__main__":
     parser.add_argument("first_json", help="first JSON to compare")
     parser.add_argument("second_json", help="second JSON to compare")
     parser.add_argument("--significant-digits", type=int, default=5, help="number of significant digits to use when comparing numeric values")
+    parser.add_argument("--exclude-paths", nargs="+", help="list of paths to exclude from consideration when performing a diff", default=["root['generated_by']['version']"])
 
     args = parser.parse_args()
 
@@ -26,6 +27,7 @@ if __name__ == "__main__":
         deepdiff.DeepDiff(
             first_json,
             second_json,
-            significant_digits=args.significant_digits
+            significant_digits=args.significant_digits,
+            exclude_paths=args.exclude_paths
         )
     )