+++ tests/builds/zika.t
@@ -7,10 +7,11 @@ Running from the test data directory allows us to use relative paths that won't 
   $ TEST_DATA_DIR="$TESTDIR/zika"
   $ mkdir -p "$TMP/out"
   $ pushd "$TEST_DATA_DIR" > /dev/null
+  $ export AUGUR="../../../bin/augur"
 
 Parse a FASTA whose defline contains metadata into separate sequence and metadata files.
 
-  $ augur parse \
+  $ ${AUGUR} parse \
   >   --sequences "data/zika.fasta" \
   >   --output-sequences "$TMP/out/sequences.fasta" \
   >   --output-metadata "$TMP/out/metadata.tsv" \
@@ -21,7 +22,7 @@ Parse a FASTA whose defline contains metadata into separate sequence and metadat
 
 Filter sequences by a minimum date and an exclusion list and only keep one sequence per country, year, and month.
 
-  $ augur filter \
+  $ ${AUGUR} filter \
   >   --sequences "results/sequences.fasta" \
   >   --metadata "results/metadata.tsv" \
   >   --exclude "config/dropped_strains.txt" \
@@ -35,7 +36,7 @@ Filter sequences by a minimum date and an exclusion list and only keep one seque
 
 Align filtered sequences to a specific reference sequence and fill any gaps.
 
-  $ augur align \
+  $ ${AUGUR} align \
   >  --sequences "results/filtered.fasta" \
   >  --reference-sequence "config/zika_outgroup.gb" \
   >  --output "$TMP/out/aligned.fasta" \
@@ -45,7 +46,7 @@ Align filtered sequences to a specific reference sequence and fill any gaps.
 
 Build a tree from the multiple sequence alignment.
 
-  $ augur tree \
+  $ ${AUGUR} tree \
   >  --alignment "results/aligned.fasta" \
   >  --output "$TMP/out/tree_raw.nwk" \
   >  --method iqtree \
@@ -54,9 +55,15 @@ Build a tree from the multiple sequence alignment.
   $ python3 "$TESTDIR/../../scripts/diff_trees.py" "results/tree_raw.nwk" "$TMP/out/tree_raw.nwk" --significant-digits 5
   {}
 
+Confirm that tree log file exists.
+
+  $ test -e "results/aligned-delim.iqtree.log"
+  $ echo $?
+  0
+
 Build a time tree from the existing tree topology, the multiple sequence alignment, and the strain metadata.
 
-  $ augur refine \
+  $ ${AUGUR} refine \
   >  --tree "results/tree_raw.nwk" \
   >  --alignment "results/aligned.fasta" \
   >  --metadata "results/metadata.tsv" \
@@ -81,29 +88,29 @@ Branch lengths and other annotations like dates are too stochastic across runs t
 
 Calculate tip frequencies from the tree.
 
-  $ augur frequencies \
+  $ ${AUGUR} frequencies \
   >  --method kde \
   >  --tree "results/tree.nwk" \
   >  --metadata "results/metadata.tsv" \
   >  --pivot-interval 3 \
   >  --output "$TMP/out/zika_tip-frequencies.json" > /dev/null
 
-  $ diff -u "auspice/zika_tip-frequencies.json" "$TMP/out/zika_tip-frequencies.json"
+  $ diff -u --ignore-matching-lines version "auspice/zika_tip-frequencies.json" "$TMP/out/zika_tip-frequencies.json"
 
 Infer ancestral sequences from the tree.
 
-  $ augur ancestral \
+  $ ${AUGUR} ancestral \
   >  --tree "results/tree.nwk" \
   >  --alignment "results/aligned.fasta" \
   >  --infer-ambiguous \
   >  --output-node-data "$TMP/out/nt_muts.json" \
   >  --inference joint > /dev/null
 
-  $ diff -u "results/nt_muts.json" "$TMP/out/nt_muts.json"
+  $ diff -u --ignore-matching-lines version "results/nt_muts.json" "$TMP/out/nt_muts.json"
 
 Infer ancestral traits from the tree.
 
-  $ augur traits \
+  $ ${AUGUR} traits \
   >  --tree "results/tree.nwk" \
   >  --weights "config/trait_weights.csv" \
   >  --metadata "results/metadata.tsv" \
@@ -123,31 +130,30 @@ See augur issue 541 (https://github.com/nextstrain/augur/issues/541) for more de
 
 Translate inferred ancestral and observed nucleotide sequences to amino acid mutations.
 
-  $ augur translate \
+  $ ${AUGUR} translate \
   >  --tree "results/tree.nwk" \
-  >  --ancestral-sequences "results/nt_muts.json" \
+  >  --ancestral-sequences "$TMP/out/nt_muts.json" \
   >  --reference-sequence "config/zika_outgroup.gb" \
   >  --output-node-data "$TMP/out/aa_muts.json" > /dev/null
 
-  $ diff -u "results/aa_muts.json" "$TMP/out/aa_muts.json"
+  $ diff -u --ignore-matching-lines version "results/aa_muts.json" "$TMP/out/aa_muts.json"
 
 Export JSON files as v1 auspice outputs.
 
-  $ augur export v1 \
+  $ ${AUGUR} export v1 \
   >  --tree "results/tree.nwk" \
   >  --metadata "results/metadata.tsv" \
-  >  --node-data "results/branch_lengths.json" \
-  >              "results/traits.json" \
-  >              "results/nt_muts.json" \
-  >              "results/aa_muts.json" \
+  >  --node-data "$TMP/out/branch_lengths.json" \
+  >              "$TMP/out/traits.json" \
+  >              "$TMP/out/nt_muts.json" \
+  >              "$TMP/out/aa_muts.json" \
   >  --colors "config/colors.tsv" \
   >  --auspice-config "config/auspice_config_v1.json" \
   >  --output-tree "$TMP/out/v1_zika_tree.json" \
   >  --output-meta "$TMP/out/v1_zika_meta.json" \
   >  --output-sequence "$TMP/out/v1_zika_seq.json" > /dev/null
 
-  $ augur validate export-v1 "$TMP/out/v1_zika_meta.json" "$TMP/out/v1_zika_tree.json" > /dev/null
-  $ diff -u "auspice/v1_zika_tree.json" "$TMP/out/v1_zika_tree.json"
+  $ ${AUGUR} validate export-v1 "$TMP/out/v1_zika_meta.json" "$TMP/out/v1_zika_tree.json" > /dev/null
 
 Compare auspice metadata files, but ignore the "updated" field since this changes with the date the export command is run.
 
@@ -155,24 +161,20 @@ Compare auspice metadata files, but ignore the "updated" field since this change
 
 Export JSON files as v2 auspice outputs.
 
-  $ augur export v2 \
+  $ ${AUGUR} export v2 \
   >  --tree "results/tree.nwk" \
   >  --metadata "results/metadata.tsv" \
-  >  --node-data "results/branch_lengths.json" \
-  >              "results/traits.json" \
-  >              "results/nt_muts.json" \
-  >              "results/aa_muts.json" \
+  >  --node-data "$TMP/out/branch_lengths.json" \
+  >              "$TMP/out/traits.json" \
+  >              "$TMP/out/nt_muts.json" \
+  >              "$TMP/out/aa_muts.json" \
   >  --colors "config/colors.tsv" \
   >  --auspice-config "config/auspice_config_v2.json" \
   >  --output "$TMP/out/v2_zika.json" \
   >  --title 'Real-time tracking of Zika virus evolution -- v2 JSON' \
   >  --panels tree map entropy frequencies > /dev/null
 
-  $ augur validate export-v2 "$TMP/out/v2_zika.json" > /dev/null
-
-Ignore the date the auspice output was "updated", but consider all other differences.
-
-  $ diff -u --ignore-matching-lines updated "auspice/v2_zika.json" "$TMP/out/v2_zika.json"
+  $ ${AUGUR} validate export-v2 "$TMP/out/v2_zika.json" > /dev/null
 
 Switch back to the original directory where testing started.