+++ tests/test_utils.py
@@ -19,12 +19,6 @@ class TestUtils:
             datetime.date(year=2000, month=1, day=31),
         )
 
-    def test_ambiguous_date_to_date_range_ambiguous_month(self):
-        assert utils.ambiguous_date_to_date_range("2000-XX-5", "%Y-%m-%d") == (
-            datetime.date(year=2000, month=1, day=5),
-            datetime.date(year=2000, month=12, day=5),
-        )
-
     def test_ambiguous_date_to_date_range_ambiguous_month_and_day(self):
         assert utils.ambiguous_date_to_date_range("2000-XX-XX", "%Y-%m-%d") == (
             datetime.date(year=2000, month=1, day=1),
@@ -106,25 +100,3 @@ class TestUtils:
         with open(drm_file, "w") as fh:
             fh.write("\n".join(drm_lines))
         assert utils.read_mask_file(drm_file) == expected_sites
-
-    def test_read_metadata_with_good_query(self, tmpdir):
-        meta_fn = str(tmpdir / "metadata.tsv")
-        meta_lines = ["strain\tlocation\tquality",
-                      "c_good\tcolorado\tgood",
-                      "c_bad\tcolorado\tbad",
-                      "n_good\tnevada\tgood"]
-        with open(meta_fn, "w") as fh:
-            fh.write("\n".join(meta_lines))
-        meta_dict, _ = utils.read_metadata(meta_fn, query='quality=="good" & location=="colorado"')
-        assert len(meta_dict) == 1
-        assert "c_good" in meta_dict
-
-    def test_read_metadata_bad_query(self, tmpdir):
-        meta_fn = str(tmpdir / "metadata.tsv")
-        meta_lines = ["strain\tlocation\tquality",
-                      "c_good\tcolorado\tgood",
-                      "n_bad\tnevada\tbad",]
-        with open(meta_fn, "w") as fh:
-            fh.write("\n".join(meta_lines))
-        with pytest.raises(SystemExit):
-            utils.read_metadata(meta_fn, query='badcol=="goodval"')